Split DockerComposeIT into a suite of DockerComposeTestPlatform, DockerComposeTestStream, DockerComposeTestTask
