+++ spring-cloud-dataflow-server/pom.xml
@@ -181,7 +181,7 @@
 				<plugins>
 					<plugin>
 						<artifactId>maven-failsafe-plugin</artifactId>
-						<version>3.0.0-M3</version>
+						<version>3.0.0-M4</version>
 						<executions>
 							<execution>
 								<goals>
